+++ module/rest-upload/__index__.ts
@@ -2,4 +2,3 @@
 
 export * from './src/decorator';
 export * from './src/config';
-export * from './src/util';
